Rewritten blit functions to work with 32bpp.

+++ src/game/Editor/Sector_Summary.cc
@@ -1223,7 +1223,7 @@ void RenderSummaryWindow()
 		MPrint(354, 18, L"Summary");
 		{ SGPVSurface::Lock l(FRAME_BUFFER);
 			SetClippingRegionAndImageWidth(l.Pitch(), 0, 0, SCREEN_WIDTH, SCREEN_HEIGHT);
-			RectangleDraw(TRUE, 350, 15, 405, 28, 0, l.Buffer<UINT16>());
+			RectangleDraw(TRUE, 350, 15, 405, 28, 0, l.Buffer<UINT32>());
 		}
 		FRAME_BUFFER->ShadowRectUsingLowPercentTable(351, 16, 404, 27);
 		if( gpCurrentSectorSummary )
@@ -1240,7 +1240,7 @@ void RenderSummaryWindow()
 		MPrint(354, 33, L"Items");
 		{ SGPVSurface::Lock l(FRAME_BUFFER);
 			SetClippingRegionAndImageWidth(l.Pitch(), 0, 0, SCREEN_WIDTH, SCREEN_HEIGHT);
-			RectangleDraw(TRUE, 350, 30, 405, 43, 0, l.Buffer<UINT16>());
+			RectangleDraw(TRUE, 350, 30, 405, 43, 0, l.Buffer<UINT32>());
 		}
 		if( gpCurrentSectorSummary )
 			/*&& gpCurrentSectorSummary->usNumItems ||
@@ -1281,7 +1281,7 @@ void RenderSummaryWindow()
 		{
 			UINT16 pos;
 			SGPVSurface::Lock l(FRAME_BUFFER);
-			UINT16* const pDestBuf = l.Buffer<UINT16>();
+			UINT32* const pDestBuf = l.Buffer<UINT32>();
 			SetClippingRegionAndImageWidth(l.Pitch(), 0, 0, SCREEN_WIDTH, SCREEN_HEIGHT);
 			for( i = 1; i <= 15; i++ )
 			{
@@ -1365,7 +1365,7 @@ void RenderSummaryWindow()
 	if( gfGlobalSummaryExists )
 	{
 		SGPVSurface::Lock l(FRAME_BUFFER);
-		UINT16* const pDestBuf = l.Buffer<UINT16>();
+		UINT32* const pDestBuf = l.Buffer<UINT32>();
 		SetClippingRegionAndImageWidth(l.Pitch(), 0, 0, SCREEN_WIDTH, SCREEN_HEIGHT);
 		//Render the grid for the map currently residing in memory (blue).
 		if( gfWorldLoaded && !gfTempFile && gsSectorX )
